refactor(api): subscriptions in isolated controller
