+++ api/src/app.module.ts
@@ -5,7 +5,7 @@ import { PrismaModule } from "nestjs-prisma";
 import { AuthModule } from "./auth/auth.module";
 import { UsersModule } from "./users/users.module";
 import { EmailModule } from "./email/email.module";
-import { CustomerModule } from "./customer/customer.module";
+import { SubscriptionsModule } from "./subscriptions/subscriptions.module";
 import configurations from "./config";
 
 @Module({
@@ -35,7 +35,7 @@ import configurations from "./config";
     AuthModule,
     UsersModule,
     EmailModule,
-    CustomerModule,
+    SubscriptionsModule,
   ],
 })
 export class AppModule {}
