--- api/src/customer/customer.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CustomerController } from './customer.controller';
-
-describe('CustomerController', () => {
-  let controller: CustomerController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [CustomerController],
-    }).compile();
-
-    controller = module.get<CustomerController>(CustomerController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
