+++ api/src/subscriptions/subscriptions.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from "@nestjs/testing";
+import { SubscriptionsController } from "./subscriptions.controller";
+
+describe("SubscriptionsController", () => {
+  let controller: SubscriptionsController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [SubscriptionsController],
+    }).compile();
+
+    controller = module.get<SubscriptionsController>(SubscriptionsController);
+  });
+
+  it("should be defined", () => {
+    expect(controller).toBeDefined();
+  });
+});
