+++ api/src/subscriptions/subscriptions.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from "@nestjs/testing";
+import { SubscriptionsService } from "./subscriptions.service";
+
+describe("SubsriptionsService", () => {
+  let service: SubscriptionsService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [SubscriptionsService],
+    }).compile();
+
+    service = module.get<SubscriptionsService>(SubscriptionsService);
+  });
+
+  it("should be defined", () => {
+    expect(service).toBeDefined();
+  });
+});
